Refactorización de tipos en servicios y handlers, creación de tipos en carpeta 'types', y finalización de configuración ESLint
- Refactorización de los tipos utilizados en los servicios y handlers para mejorar la consistencia y claridad del código.
- Creación de tipos específicos en la carpeta 'types' para estructuras de datos y objetos complejos.
- Implementación y finalización de la configuración de ESLint en todo el proyecto para asegurar cumplimiento de normas de estilo y buenas prácticas.

+++ src/libs/services/handler/delete.ts
@@ -1,30 +1,34 @@
-import { NextApiRequest, NextApiResponse } from "next";
+import { NextApiResponse } from 'next'
 
-import { dataService } from "@/libs/services/dataService";
-import { messageCRUD } from "@/libs/message";
+import { PrismaDelegate } from '@/types/prismaDelegate'
+import { dataService } from '@/libs/services/dataService'
+import { messageCRUD } from '@/libs/message'
+import { RequestProps } from '@/types/RequestProps'
 
 /**
  * Handles the DELETE request for deleting a resource.
  *
- * @param {NextApiRequest} req - The request object.
+ * @param {Pick<RequestProps, 'method' | 'body'>} req - The request object containing the method and body.
  * @param {NextApiResponse} res - The response object.
- * @param {any} model - The model for the resource.
+ * @param {PrismaDelegate<unknown>} model - The PrismaDelegate representing the model for the resource.
  * @return {Promise<void>} - Returns a promise that resolves to void.
  */
 export default async function handler(
-	req: NextApiRequest,
-	res: NextApiResponse,
-	model:any
+  req: Pick<RequestProps, 'method' | 'body'>,
+  res: NextApiResponse,
+  model: PrismaDelegate<unknown>,
 ) {
-	if (req.method !== "DELETE") {
-		res.setHeader("Allow", ["DELETE"]);
-		return res.status(405).end(`Metodo ${req.method} no permitido`);
-	}
-	try {
-		const { id } = req.body;
-		await dataService.destroy(model, id);
-		return res.status(204).end();
-	} catch (error) {
-		return res.status(500).json({ error: messageCRUD.error.delete });
-	}
+  if (req.method !== 'DELETE') {
+    res.setHeader('Allow', ['DELETE'])
+    return res.status(405).end(`Metodo ${req.method} no permitido`)
+  }
+  try {
+    const { id } = req.body
+    if (id) {
+      await dataService.destroy(model, id)
+      return res.status(204).end()
+    }
+  } catch (error) {
+    return res.status(500).json({ error: messageCRUD.error.delete })
+  }
 }
